--- a/package.json
+++ b/package.json
@@ -15,6 +15,7 @@
     "react": "^18",
     "react-dom": "^18",
     "react-hook-form": "^7.52.1",
+    "react-virtuoso": "^4.7.11",
     "zustand": "^4.5.4"
   },
   "devDependencies": {