--- a/src/app/(home)/page.tsx
+++ b/src/app/(home)/page.tsx
@@ -2,6 +2,7 @@ import Header from '../../shared/components/common/Header'
 import Footer from '../../shared/components/common/Footer'
 import { getHoneyPlaces } from '../api/getHoneyPlace'
 import HoneyPlaceListClinet from '@/features/honeyPlaceList/HoneyPlaceListClient'
+import mockPlaces from '@/app/api/windowingTestData'
 
 export default function Home() {
   return (
@@ -14,6 +15,7 @@ export default function Home() {
 }
 
 const HoneyPlaceListServer = async () => {
-  const places = await getHoneyPlaces()
-  return <HoneyPlaceListClinet initialPlaces={places} />
+  //const places = await getHoneyPlaces()
+
+  return <HoneyPlaceListClinet initialPlaces={mockPlaces} />
 }