--- a/src/features/honeyPlaceList/HoneyPlaceCard.tsx
+++ b/src/features/honeyPlaceList/HoneyPlaceCard.tsx
@@ -2,10 +2,10 @@ import React from 'react'
 import MarkerIcon from '@/assets/icons/marker.svg'
 import { HoneyPlace } from '@/interfaces/IPlace'
 import truncateAddress from '@/utils/truncateAddress'
+import Image from 'next/image'
 
 const cardContainer = 'flex flex-col w-[154px] h-[301px] gap-[10px] rounded-[4px] overflow-hidden'
 const imageContainer = 'w-[150px] h-[200px] bg-slate-100'
-const imageClass = 'w-full h-full object-cover'
 const textContainer = 'px-[2px]'
 const titleClass =
   'text-[14px] font-semibold leading-[22px] text-left text-[#1c1c1c] whitespace-nowrap overflow-hidden text-ellipsis'
@@ -18,7 +18,15 @@ const HoneyPlaceCard = ({ place }: { place: HoneyPlace }) => {
   return (
     <div className={cardContainer}>
       <div className={imageContainer}>
-        <img className={imageClass} src={place?.images[0]} alt={place?.name} />
+        {/* <img className={imageClass} src={place?.images[0]} alt={place?.name} /> */}
+        <Image
+          className='w-full h-full object-cover'
+          src={place?.images[0]}
+          alt={place?.name}
+          width={150}
+          height={200}
+          loading='eager'
+        />
       </div>
       <div className={textContainer}>
         <h3 className={titleClass}>{place?.name}</h3>