… public/testimage.PNG