# Next의 Image태그와 windowing 처리

## 1. img 태그 -> Next Image 태그 사용
홈 화면에서 이미지가 담긴 게시물을 표시하고 있습니다. 이미지 최적화를 위해 Next의 Image 태그를 사용

변경 전:
리소스: 19.3MB
DOMContentLoaded: 363ms
로드: 1초

변경 후:
리소스: 12.1MB
DOMContentLoaded: 140ms
로드: 798ms

게시물 수가 10개였기에 큰 차이는 없지만, 성능 향상이 확인됨

## 2. 가상화 리스트 적용
웹 애플리케이션의 성능 최적화를 위해 목업 데이터를 사용하여 가상화 리스트를 테스트.

#### 문제점
목업 데이터 1000개를 한 번에 불러와 화면에 표시했을 때 페이지네이션 없이 모든 데이터를 렌더링하면 로드 시간이 길고, React 프로파일러에서 성능 문제가 나타남.
무한스크롤을 다른 프로젝트에서 적용해봤는데, 무한 스크롤 페이지네이션을 계속 사용하다보면 요소가 쌓이니 그 상황을 가정하여 가상화를 적용해 성능 개선을 시도

### 가상화 적용
react-vertuoso 라이브러리를 사용하여 windowing 적용.

### 적용 전후 비교
#### 네트워크 성능:
DOMContentLoaded: 1.11초 ->  313ms
로드: 1.63초 -> 832ms

#### React 프로파일러 성능:
Render: 382.5ms -> 65.7ms
Layout effects: 189.8ms -> 0.4ms
Passive effects: 9.7ms -> 3.7ms

### windowing 적용 시 문제와 개선
처음에 `react-window` 를 사용했고 스크롤 분리와 css 적용문제로 라이브러리를 `react-virtuoso`로 교체

#### 스크롤 문제
- `react-window` 사용 시 내부 스크롤이 메인 브라우저 스크롤과 분리되는 문제 발생
- `react-virtuoso` 라이브러리로 변경 → `useWindowScroll` 속성으로 해결.

#### CSS 문제
- 홈 화면의 게시물 리스트가 2열의 그리드 UI로 구성되어 있음
- `VirtuosoGrid` 태그와 부모 태그의 Tailwind CSS 일부 적용이 안되는 문제 발생
- 외부 라이브러리에 tailwind css가 제대로 적용이 안되었을 수 있음
- 우선순위가 더 높은 인라인 스타일로 작성하니 해결됨